refactor(i18n): move locale files to specific directories
Prepare for sharding up of locale resources by moving each locale into
its own directory. Use this as the base to manage future growth and
expansion of locale resources to cover most text in FormSG.

No code change.
